Python/My Tools/requirements.txt
OpenAI
load_dotenv
truststore